+++ server/misc/movement.cc
@@ -1652,7 +1652,7 @@ int TBeing::genericMovedIntoRoom(TRoom *rp, int was_in,
       continue;
 
     if (was_in != -1) {
-      rc = tmons->checkSpec(this, CMD_MOB_MOVED_INTO_ROOM, "", (TObj *) was_in);
+      rc = tmons->checkSpec(this, CMD_MOB_MOVED_INTO_ROOM, "", (TObj *)(long) was_in);
       if (rc) {
         // if TRUE, was prevented from entering.  
         // will have been put back into original room
